+++ apps/api/src/app/admin/admin.service.ts
@@ -331,19 +331,35 @@ export class AdminService {
     symbol,
     symbolMapping
   }: Prisma.SymbolProfileUpdateInput & UniqueAsset) {
-    await this.symbolProfileService.updateSymbolProfile({
-      assetClass,
-      assetSubClass,
-      comment,
-      countries,
-      currency,
-      dataSource,
-      name,
-      scraperConfiguration,
-      sectors,
-      symbol,
-      symbolMapping
-    });
+    const updatedSymbolProfile: Prisma.SymbolProfileUpdateInput & UniqueAsset =
+      {
+        assetClass,
+        assetSubClass,
+        comment,
+        countries,
+        currency,
+        dataSource,
+        scraperConfiguration,
+        sectors,
+        symbol,
+        symbolMapping,
+        ...(dataSource === 'MANUAL'
+          ? { name }
+          : {
+              SymbolProfileOverrides: {
+                upsert: {
+                  create: {
+                    name: name as string
+                  },
+                  update: {
+                    name: name as string
+                  }
+                }
+              }
+            })
+      };
+
+    await this.symbolProfileService.updateSymbolProfile(updatedSymbolProfile);
 
     const [symbolProfile] = await this.symbolProfileService.getSymbolProfiles([
       {